PAAC-154: Consolidated package Utilities into calculators.internal.Utilities. Moved CalculatorFactory into calculators.results. Updated calculators in calculators.results to be protected where not necessary to expose outside the package.
