+++ app/calculators/results/BasicCalculator.scala
@@ -17,13 +17,12 @@
 package calculators.results
 
 import models._
-import calculators.AllowanceCalculator 
-import calculators.internal._ 
+import calculators.internal.BasicAllowanceCalculator
 
 /**
  Base trait for calculators that calculate results and create summary objects
  */
-protected trait BasicCalculator extends AllowanceCalculator {
+protected trait BasicCalculator extends FactoryCalculator {
   protected def getAnnualAllowanceInPounds: Long
 
   def allowance(): Long = getAnnualAllowanceInPounds * 100L
