+++ app/calculators/results/Year2015Period1Calculator.scala
@@ -17,14 +17,17 @@
 package calculators.results
 
 import models._
-import calculators.internal._
+import calculators.internal.PeriodCalculator
+import calculators.internal.ExtendedSummaryCalculator
 
 /**
   Calculator for 2015, period 1 (up to 8th July 2015)
 */
-object Year2015Period1Calculator extends ExtendedCalculator {
+protected trait Year2015Period1Calculator extends ExtendedCalculator {
   protected def getAnnualAllowanceInPounds: Long = 80000L
   protected def getCalculator(implicit previousPeriods:Seq[TaxYearResults], contribution: Contribution): ExtendedSummaryCalculator = 
     PeriodCalculator(getAnnualAllowanceInPounds)
   def isSupported(contribution:Contribution): Boolean = contribution.isPeriod1() && !contribution.isEmpty
 }
+
+protected object Year2015Period1Calculator extends Year2015Period1Calculator
